Extract SceneController superclass which is subclassed by both scene controllers.
